Clean up and split further Product Collection E2Es for CI eff.
