+++ plugins/woocommerce-blocks/tests/e2e/tests/product-collection/compatibility-layer.block_theme.spec.ts
@@ -86,35 +86,33 @@ const test = base.extend< { pageObject: ProductCollectionPage } >( {
 	},
 } );
 
-test.describe( 'Compatibility Layer with Product Collection block', () => {
-	test.describe( 'Product Archive with Product Collection block', () => {
-		test.beforeEach( async ( { pageObject, requestUtils } ) => {
-			await requestUtils.activatePlugin(
-				'woocommerce-blocks-test-product-collection-compatibility-layer'
-			);
-			await pageObject.goToProductCatalogFrontend();
-		} );
-
-		for ( const scenario of singleOccurrenceScenarios ) {
-			test( `${ scenario.title } is attached to the page`, async ( {
-				pageObject,
-			} ) => {
-				const hooks = pageObject.locateByTestId( scenario.dataTestId );
-
-				await expect( hooks ).toHaveCount( scenario.amount );
-				await expect( hooks ).toHaveText( scenario.content );
-			} );
-		}
-
-		for ( const scenario of multipleOccurrenceScenarios ) {
-			test( `${ scenario.title } is attached to the page`, async ( {
-				pageObject,
-			} ) => {
-				const hooks = pageObject.locateByTestId( scenario.dataTestId );
-
-				await expect( hooks ).toHaveCount( scenario.amount );
-				await expect( hooks.first() ).toHaveText( scenario.content );
-			} );
-		}
+test.describe( 'Product Collection: Compatibility Layer', () => {
+	test.beforeEach( async ( { pageObject, requestUtils } ) => {
+		await requestUtils.activatePlugin(
+			'woocommerce-blocks-test-product-collection-compatibility-layer'
+		);
+		await pageObject.goToProductCatalogFrontend();
 	} );
+
+	for ( const scenario of singleOccurrenceScenarios ) {
+		test( `${ scenario.title } is attached to the page`, async ( {
+			pageObject,
+		} ) => {
+			const hooks = pageObject.locateByTestId( scenario.dataTestId );
+
+			await expect( hooks ).toHaveCount( scenario.amount );
+			await expect( hooks ).toHaveText( scenario.content );
+		} );
+	}
+
+	for ( const scenario of multipleOccurrenceScenarios ) {
+		test( `${ scenario.title } is attached to the page`, async ( {
+			pageObject,
+		} ) => {
+			const hooks = pageObject.locateByTestId( scenario.dataTestId );
+
+			await expect( hooks ).toHaveCount( scenario.amount );
+			await expect( hooks.first() ).toHaveText( scenario.content );
+		} );
+	}
 } );
